// src/app/models/vw-apvendor-master.interface.ts
export interface IVwApvendorMaster {
	vactiv: string;
	vnumb: number;
	vname: string;
	// vaddr1: string;
	// vaddr2: string;
	// vcity: string;
	// vstate: string;
	// vzip1: number;
	// vzip2: number;
	// vphone: number;
	// vshnam: string;
	// vpaytp: string;
	// vstmt: string;
	// vrelat: string;
	// vdistr: string;
	// vutlty: string;
	// vtintp: number;
	// vgrowr: string;
	// vgidno: number;
	// vdscnt: number;
	// vterms: number;
	// vnegby: string;
	// vckdaj: number;
	// vckisu: number;
	// vcytdv: number;
	// vcytdg: number;
	// vlytdv: number;
	// vlytdg: number;
	// vpytdv: number;
	// vpytdg: number;
	// vdsytd: number;
	// vrecck: number;
	// v1099t: number;
	// vbuwth: string;
	// vbwpct: number;
	// vocrgd: number;
	// vocrjd: number;
	// vactgd: number;
	// vactjd: number;
}